refactor more code catalyst code
